RavenDB-3765 fix

+++ Raven.Abstractions/Data/PeriodicBackupSetup.cs
@@ -55,12 +55,12 @@ namespace Raven.Abstractions.Data
 		/// <summary>
 		/// Interval between incremental backups in milliseconds. If set to null or 0 then incremental periodic export will be disabled.
 		/// </summary>
-		public int? IntervalMilliseconds { get; set; }
+		public long? IntervalMilliseconds { get; set; }
 
 		/// <summary>
 		/// Interval between full backups in milliseconds. If set to null or 0 then full periodic export will be disabled.
 		/// </summary>
-        public int? FullBackupIntervalMilliseconds { get; set; }
+        public long? FullBackupIntervalMilliseconds { get; set; }
 
 		protected bool Equals(PeriodicExportSetup other)
 		{
@@ -90,8 +90,8 @@ namespace Raven.Abstractions.Data
 				hashCode = (hashCode*397) ^ (AwsRegionEndpoint != null ? AwsRegionEndpoint.GetHashCode() : 0);
 				hashCode = (hashCode*397) ^ (AzureStorageContainer != null ? AzureStorageContainer.GetHashCode() : 0);
 				hashCode = (hashCode*397) ^ (LocalFolderName != null ? LocalFolderName.GetHashCode() : 0);
-				hashCode = (hashCode*397) ^ IntervalMilliseconds.GetValueOrDefault();
-			    hashCode = (hashCode*397) ^ FullBackupIntervalMilliseconds.GetValueOrDefault();
+				hashCode = (hashCode*397) ^ (int)IntervalMilliseconds.GetValueOrDefault();
+			    hashCode = (hashCode*397) ^ (int)FullBackupIntervalMilliseconds.GetValueOrDefault();
 				return hashCode;
 			}
 		}
